rename to avoid ModuleNotFound error (#1707)
