+++ AzureMonitorAgent/ama_tst/modules/syslog_tst/syslog.py
